# Conflicts:
#	config/nuls.beta.ncf
#	config/nuls.ncf
#	module.ncf
#	module/nuls-cores/src/main/java/io/nuls/account/util/validator/TxValidator.java
#	module/nuls-cores/src/main/java/io/nuls/common/ConfigBean.java
#	module/nuls-cores/src/main/java/io/nuls/consensus/utils/manager/ConsensusManager.java
#	module/nuls-cores/src/main/java/io/nuls/contract/config/ContractContext.java
#	module/nuls-cores/src/main/java/io/nuls/contract/helper/ContractHelper.java
#	module/nuls-cores/src/main/java/io/nuls/contract/model/dto/ContractResultDto.java
#	module/nuls-cores/src/main/java/io/nuls/contract/util/ContractUtil.java
#	module/nuls-cores/src/main/java/io/nuls/contract/validator/CallContractTxValidator.java
#	module/nuls-cores/src/main/java/io/nuls/contract/validator/CreateContractTxValidator.java
#	module/nuls-cores/src/main/java/io/nuls/protocol/service/impl/ProtocolServiceImpl.java
#	module/nuls-cores/src/main/java/io/nuls/transaction/service/impl/TxServiceImpl.java
#	module/nuls-cores/src/test/java/io/nuls/account/service/AccountServiceTest.java
#	module/nuls-cores/src/test/java/io/nuls/contract/mock/basetest/MockBase.java
#	module/nuls-cores/src/test/java/io/nuls/contract/mock/storagestructure/LoadLargeContractTest.java
#	module/nuls-cores/src/test/java/io/nuls/contract/tx/contractvm/ContractVmV8SendTxTest.java
#	module/nuls-cores/src/test/java/io/nuls/contract/tx/nrc20/ContractNRC20TokenSendTxTest.java

+++ module/nuls-cores/src/main/java/io/nuls/protocol/storage/ProtocolVersionStorageService.java
@@ -46,7 +46,7 @@ public interface ProtocolVersionStorageService {
      */
     List<ProtocolVersionPo> getList(int chainId);
 
-    boolean saveCurrentProtocolVersionCount(int chainId, int currentProtocolVersionCount);
+    boolean saveCurrentProtocolVersionCount(int chainId, long currentProtocolVersionCount);
 
-    int getCurrentProtocolVersionCount(int chainId);
+    long getCurrentProtocolVersionCount(int chainId);
 }
